Changed number from 191 to 032 + modified the goal
Changed to goal to modifying a received string to start and end with quotation marks. Added tests to verify this !
