+++ Ellabit/Challenges/Challenges.cs
@@ -36,6 +36,7 @@ namespace Ellabit.Challenges
             challenges.Add(new Challenge029Correctthemistakes());
             challenges.Add(new Challenge030Fliptheintegerboolean());
             challenges.Add(new Challenge031Dividesevenly());
+            challenges.Add(new Challenge032StringEscape());
             challenges.Add(new Challenge033Profitablegamble());
             challenges.Add(new Challenge034Multipleof100());
             challenges.Add(new Challenge035Isittimeformilkandcookies());
@@ -187,7 +188,6 @@ namespace Ellabit.Challenges
             challenges.Add(new Challenge188UTF8bomtextencoding());
             challenges.Add(new Challenge189Isthenumbersymmetrical());
             challenges.Add(new Challenge190Maskifythestring());
-            challenges.Add(new Challenge191StringEscape());
             challenges.Add(new ChallengeColor001());
             return challenges;
         }
